Refactor NearbyManager to helper package
