+++ app/src/main/java/com/andrerinas/openheadunit/connection/CommManager.kt
@@ -25,7 +25,7 @@ import java.net.Socket
 import android.view.KeyEvent
 import com.andrerinas.openheadunit.aap.protocol.messages.TouchEvent
 import com.andrerinas.openheadunit.aap.protocol.proto.Input.TouchEvent.PointerAction
-import com.andrerinas.openheadunit.connection.wifi.NearbySocket
+import com.andrerinas.openheadunit.connection.wifi.modes.helper.NearbySocket
 
 /**
  * Central connection and transport lifecycle manager.
